[Ui]: Apply responsive design to pages 'login', 'signup', 'user'; Using the 'viewport' meta tag to control layout on mobile browsers. [Temp]: 'lobby' page read query params 'name' and 'page'.
