+++ client/lib/avalon.css
@@ -89,7 +89,7 @@ h3 small {
   opacity: 1 !important;
 }
 
-ul.bar_tabs>li a,
+ul.bar_tabs > li a,
 .x_panel {
   padding: 10px 8px;
 }
@@ -110,6 +110,14 @@ ul.bar_tabs>li a,
   ul.bar_tabs > li {
     margin-left: 0px;
   }
+}
+
+.login_wrapper {
+  margin-top: 60px;
+}
+
+.login_content {
+  min-width: 225px;
 }
 
 /*
@@ -159,7 +167,6 @@ footer {
   background-color: rgba(255,255,255,0.9);
   position: inherit;
   padding: 10px 20px;
-  margin: 20px;
 }
 
 .page-title .title_left h3 {
@@ -454,7 +461,43 @@ button.avalon-card-disable {
 
 .avalon-chart {
   width: 100%;
-  height: 280px;
+  height: 160px;
+}
+
+.profile_left .avalon-chart {
+  height: 120px;
+  margin: 0px auto;
+}
+
+@media (min-width: 480px) {
+  .avalon-chart {
+    height: 200px;
+  }
+
+  .profile_left .avalon-chart {
+    height: 140px;
+  }
+}
+
+@media (min-width: 624px) {
+  .avalon-chart {
+    height: 240px;
+  }
+
+  .profile_left .avalon-chart {
+    height: 160px;
+  }
+}
+
+@media (min-width: 768px) {
+  .avalon-chart {
+    height: 280px;
+  }
+
+  .profile_left .avalon-chart {
+    width: 80%;
+    height: 180px;
+  }
 }
 
 tspan {
