+++ client/main.html
@@ -1,5 +1,6 @@
 <head>
   <title>Avalon</title>
+  <meta name=viewport content="width=device-width, initial-scale=1.0, minimum-scale=0.5 maximum-scale=1.0">
 </head>
 <body>
     <div id="app" class="right-col" role="main"></div>
