+++ imports/api/groups/server/publications.js
@@ -2,7 +2,19 @@ import { Meteor } from 'meteor/meteor';
 import { Groups } from '../groups.jsx';
 import { check } from 'meteor/check';
 
-Meteor.publish('groups.findAll', () => Groups.find({}, { fields: Groups.publicFieldsWhenFindAll }));
+Meteor.publish('groups.findAll', (name, page) => {
+  check(name, String);
+  check(page, Number);
+  const GROUPS_PER_PAGE = Groups.find().count();
+  const skippedGroupsCount = (page - 1) * GROUPS_PER_PAGE;
+  const selector = { name: { $regex: name } };
+  if (skippedGroupsCount > 0) {
+    const lastGroup = Groups.find(selector, { sort: { createdAt: -1 }, limit: skippedGroupsCount }).fetch().pop();
+    selector.createdAt = { $lt: lastGroup.createdAt };
+  }
+  return Groups.find(selector, { fields: Groups.publicFieldsWhenFindAll, sort: { createdAt: -1 }, limit: GROUPS_PER_PAGE });
+});
+
 Meteor.publish('groups.findOne', id => {
   check(id, String);
   return Groups.find({ _id: id }, { fields: Groups.publicFieldsWhenFindOne });
